Update OOMA.md
- Minor updates to save characters. 
- New beginning sections; Directions, Considerations. (To be appended to and revised.)

Was going to be more, but changed mind. (Considered inserting a [OOMA_NAME] variable field into the template as to prompt the user to focus on that area and/or just update the OOMA name in the CusIn Master template as I've done. But ultimately, that seems to tedious and a simple enough fix with your current version. 

*Consider future addition of a footnote/disclaimer about being a work in progress, and provide an open invitation to submit forks, and or public invite to the Prompt Engineering community with invite link? (Stuff like that?)